Move `internal` package to root.
