+++ src/ax/bus/context_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	. "github.com/jmalloc/ax/src/ax/bus"
-	"github.com/jmalloc/ax/src/ax/internal/bustest"
-	"github.com/jmalloc/ax/src/ax/internal/messagetest"
+	"github.com/jmalloc/ax/src/internal/bustest"
+	"github.com/jmalloc/ax/src/internal/messagetest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
