+++ src/ax/envelope_test.go
@@ -2,7 +2,7 @@ package ax_test
 
 import (
 	. "github.com/jmalloc/ax/src/ax"
-	"github.com/jmalloc/ax/src/ax/internal/messagetest"
+	"github.com/jmalloc/ax/src/internal/messagetest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	uuid "github.com/satori/go.uuid"
