+++ src/ax/marshaling/message_test.go
@@ -2,8 +2,8 @@ package marshaling_test
 
 import (
 	"github.com/golang/protobuf/proto"
-	"github.com/jmalloc/ax/src/ax/internal/messagetest"
 	. "github.com/jmalloc/ax/src/ax/marshaling"
+	"github.com/jmalloc/ax/src/internal/messagetest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
