+++ src/ax/persistence/injector_test.go
@@ -4,9 +4,9 @@ import (
 	"context"
 
 	"github.com/jmalloc/ax/src/ax/bus"
-	"github.com/jmalloc/ax/src/ax/internal/bustest"
-	"github.com/jmalloc/ax/src/ax/internal/persistencetest"
 	. "github.com/jmalloc/ax/src/ax/persistence"
+	"github.com/jmalloc/ax/src/internal/bustest"
+	"github.com/jmalloc/ax/src/internal/persistencetest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
