+++ src/ax/type_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 
 	. "github.com/jmalloc/ax/src/ax"
-	"github.com/jmalloc/ax/src/ax/internal/messagetest"
+	"github.com/jmalloc/ax/src/internal/messagetest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
